+++ client/src/components/Navigation/Login/NavLogin.vue
@@ -6,7 +6,7 @@
     class="testing-login-dropdown is-open-on-mobile"
   >
     <template #content>
-      <LoginEntry />
+      <LoginForm />
     </template>
     <template #actions>
       <a
@@ -27,11 +27,11 @@
 <script>
 // Components
 import Dropdown from '../_NavItems/NavDropdown'
-import LoginEntry from './NavLoginEntry'
+import LoginForm from '../../Login/LoginForm.vue'
 
 export default {
   name: 'MenuStores',
-  components: { Dropdown, LoginEntry },
+  components: { Dropdown, LoginForm },
 }
 </script>
 <style lang="scss" scoped>