+++ src/Modules/Login/Login.js
@@ -2,12 +2,12 @@ import '@/core'
 import '@/globals'
 import './Login.css'
 import { vueApply, vueRegister } from '@/vue'
-import LoginForm from './components/LoginForm.vue'
+import LoginPage from '@/views/pages/Login/LoginPage.vue'
 
 vueRegister({
-  LoginForm,
+  LoginPage,
 })
-const selector = '#login-form'
+const selector = '#login-page'
 const elements = document.querySelectorAll(selector)
 if (Array.from(elements)?.length) {
   vueApply(selector)